requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
alembic
bcrypt==4.0.1
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
loguru
email-validator
pytest-cov
python-multipart
